Exploration - Feature tree / converting from raw to tree Node
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**converting from raw tree Node**

Making sure we can handle tree nodes and build a linked draft content state

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1501

Differential Revision: D6319768

fbshipit-source-id: 566fe12aa2b8c0424d425e8c91b4560ad1dcd8e1