Exploration - Feature tree / Convert from html tree data support
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Convert from html tree data support**

This PR adds support for converting from HTML to tree data, as a follow up we will then hook the paste processor with this too

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1518

Differential Revision: D6372235

fbshipit-source-id: fdfe81efcb81e31aa4bb61f5cfe7b21bf11317a4