Exploration - Feature tree / Wiring up tree adapter on the raw
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Wiring up tree adapter on the raw**

Making sure when the darkfeature flag is enabled/disabled we can handle both data formats to guarantee some degree of migration/backwards compatibility

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1532

Differential Revision: D6448468

fbshipit-source-id: 494bb1412e2366730c23b6a5cc5521d0199939a5