Update Jest to 21.3.0-beta.13
Summary: This diff updates Jest to the latest version, and fixes some tests that are breaking. Notably tests overriding the `value` property of an `<input type="file" />`, and tests related to `<a>` URL parsing. I also removed a test that was testing a non-existing file (https://our.intern.facebook.com/intern/testinfra/testdetail/113778710).

Reviewed By: cpojer

Differential Revision: D6508625

fbshipit-source-id: 7f164d8e808c75456123fc64b80a2546ad7d8321