Warn if 'moveSelectionForward/Backward' called with non-collapsed selection
Summary:
Noticed that these helper methods are designed with the assumption they will
only be passed a collapsed selection, but I don't see anything enforcing that
requirement in the code. Since we will be tweaking these helpers and refactoring
things, want to make the code more sturdy and enforce this constraint.

Starting with a 'warning' in case something relies on broken behavior here, but
as a follow-up we should make this an invariant.

Reviewed By: vdurmont, mitermayer

Differential Revision: D9034617

fbshipit-source-id: 378390c308d1ab2657bea7a5cc9cae060e515658