5/n Disable forward delete across blocks when nested blocks are enabled
Summary:
We are disabling forward delete across blocks for now because it is rarely used
compared to backspace. We expect that people may not even notice, but if/when
they do then part of the roll-out for nested blocks will be to implement this
feature.

For more details about our ideal implementation of the forward-delete behavior
across nested blocks, see https://fb.quip.com/tybJAJoCmiRL

Reviewed By: mitermayer

Differential Revision: D9186965

fbshipit-source-id: a2abb17dd8a77949616d7e057da2476c534f2a77