added favicon (#1871)
Summary:
Fixed #1857 by adding a favicon. The color scheme of the icon needs to be thought about though since it is not visible when the tab is active because of the white color.

On chrome:
![image](https://user-images.githubusercontent.com/1450819/45538271-39089f80-b824-11e8-8553-5902b11a53bb.png)

Looks okay when the tab is inactive, but when it is active, the logo is not visible:
![image](https://user-images.githubusercontent.com/1450819/45538295-4b82d900-b824-11e8-9b27-339dc2c797d3.png)
Pull Request resolved: https://github.com/facebook/draft-js/pull/1871

Differential Revision: D9829164

fbshipit-source-id: 99c7f4d805b7da2713c3596992b39c125b63b221