Specify correct type of joinClasses
Summary:
The type of `joinClasses` in fbjs 1.0 which draft-js open-source uses is incorrect (it is correct in www), specify the correct type when importing.

I would have thought that I would have to cast the import to `any` before casting it to a different type, but apprently this works (tested in cloned github repo).

Reviewed By: claudiopro

Differential Revision: D14318804

fbshipit-source-id: e4fd34ca5efda6fd9be681a49905d9e9fdb9e3ae