Fix the issue of draft JS does not do copy and paste correctly with custom entities.
Summary:
details are here: https://github.com/facebook/draft-js/pull/1784
in the comment, nivedita mentioned we need to test this out for intern editor
hence not merging the PR.

since that comment was made back in Jan, so it's probably gonna take a while
for this to be merged into FB codebase. In the mean time this is blocking
CMS Editor's embeded CMS use case. So I figured we should do a branching
to merge the fix just for CMS Editor.

Unfortunately, the logic is pretty deep in the DraftJS stack, so I had to make
a couple changes to pass down the copy paste behavior.

Differential Revision: D14626028

fbshipit-source-id: 4826ea33abf2618835841200af81e10d707aa23b