Conditionally invokes URI.tryParseURI in OSS code
Summary:
Conditionally invokes URI.tryParseURI in OSS code because it's not available in fbjs.

I am on the fence on this one, we need it internally but we can't break OSS consumers of Draft. I think the optional chaining operator with a fallback is a pretty conservative option, however it does look clowny af.

Differential Revision: D22722736

fbshipit-source-id: d8630899b88bd0961ed2e8be86ab25d4369d40d6